Refactor MySQL PersistenceLayer (#1442)
refactoring to multiple files to improve readability
